feat: add hw-js-7
