+++ game/index.js
@@ -1,0 +1,10 @@
+import { myAttack, enemyAttack } from './actionsCharacter.js';
+import { createPlayer } from './createElementsDocument.js';
+import { resultWins } from './outputInfo.js';
+
+export {
+  myAttack,
+  enemyAttack,
+  createPlayer,
+  resultWins,
+};
